refactor: 설문지 엔티티 수정 및 클래스 구조 재구성
