--- src/main/java/com/project/always/security/oauth/entity/ProfileUrl.java
@@ -1,12 +1,0 @@
-package com.project.always.security.oauth.entity;
-
-import javax.persistence.Embeddable;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@Embeddable
-@NoArgsConstructor
-public class ProfileUrl {
-
-}
