+++ src/main/java/com/project/always/security/oauth/entity/UserMbti.java
@@ -1,4 +1,4 @@
-package com.project.always.usermbti.domain;
+package com.project.always.security.oauth.entity;
 
 import static javax.persistence.FetchType.LAZY;
 
@@ -6,6 +6,7 @@ import com.project.always.mbti.domain.Mbti;
 import com.project.always.security.oauth.entity.User;
 import com.project.always.survey.domain.Survey;
 import com.project.always.utils.BaseEntity;
+import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
@@ -13,16 +14,18 @@ import javax.persistence.Id;
 import javax.persistence.JoinColumn;
 import javax.persistence.ManyToOne;
 import lombok.AccessLevel;
+import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @Entity
-public class UserMbti extends BaseEntity {
+public class UserMbti {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Column(name = "user_mbti_id")
     private Long id;
 
     @ManyToOne(fetch = LAZY)
@@ -33,7 +36,8 @@ public class UserMbti extends BaseEntity {
     @JoinColumn(name = "mbti_id")
     private Mbti mbti;
 
-    public UserMbti(User user, Mbti mbti) {
+    @Builder
+    private UserMbti(User user, Mbti mbti) {
         this.user = user;
         this.mbti = mbti;
     }
