+++ src/main/java/com/project/always/security/oauth/entity/UserSurvey.java
@@ -1,10 +1,11 @@
-package com.project.always.usersurvey.domain;
+package com.project.always.security.oauth.entity;
 
 import static javax.persistence.FetchType.LAZY;
 
 import com.project.always.security.oauth.entity.User;
 import com.project.always.survey.domain.Survey;
 import com.project.always.utils.BaseEntity;
+import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
@@ -12,16 +13,18 @@ import javax.persistence.Id;
 import javax.persistence.JoinColumn;
 import javax.persistence.ManyToOne;
 import lombok.AccessLevel;
+import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @Entity
-public class UserSurvey extends BaseEntity {
+public class UserSurvey{
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Column(name = "user_survey_id")
     private Long id;
 
     @ManyToOne(fetch = LAZY)
@@ -34,9 +37,11 @@ public class UserSurvey extends BaseEntity {
 
     private int select_option;
 
-    public UserSurvey(User user, Survey survey) {
+    @Builder
+    private UserSurvey(final User user,final Survey survey, final int select_option) {
         this.user = user;
         this.survey = survey;
+        this.select_option = select_option;
     }
 }
