+++ src/main/java/com/project/always/survey/domain/Survey.java
@@ -6,6 +6,7 @@ import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
 import lombok.AccessLevel;
+import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
@@ -24,4 +25,11 @@ public class Survey {
     private String option1;
 
     private String option2;
+
+    @Builder
+    private Survey(final String questionText, final String option1, final String option2) {
+        this.questionText = questionText;
+        this.option1 = option1;
+        this.option2 = option2;
+    }
 }
