--- src/main/java/com/project/always/survey/domain/SurveyRepository.java
@@ -1,7 +1,0 @@
-package com.project.always.survey.domain;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface SurveyRepository extends JpaRepository<Survey, Long> {
-
-}
