+++ src/main/java/com/project/always/survey/repository/SurveyRepository.java
@@ -1,0 +1,10 @@
+package com.project.always.survey.repository;
+
+import com.project.always.survey.domain.Survey;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface SurveyRepository extends JpaRepository<Survey, Long> {
+
+}
