--- src/main/java/com/project/always/usersurvey/domain/UserSurveyRepository.java
@@ -1,8 +1,0 @@
-package com.project.always.usersurvey.domain;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface UserSurveyRepository extends JpaRepository<UserSurvey, Long> {
-}
