+++ src/test/java/com/project/always/security/oauth/controller/OauthControllerApiTest.java
@@ -15,24 +15,14 @@ import static org.springframework.restdocs.payload.PayloadDocumentation.response
 import static org.springframework.restdocs.request.RequestDocumentation.requestParts;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
-import com.epages.restdocs.apispec.MockMvcRestDocumentationWrapper;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.project.always.AcceptanceTest;
-import com.project.always.controller.api.BaseControllerTest;
 import com.project.always.infrastructure.WithMockOAuth2User;
-import com.project.always.security.oauth.dto.UserNameRequestDto;
-import com.project.always.security.oauth.dto.UserNameResponseDto;
-import com.project.always.security.oauth.dto.UserResponseDto;
-import com.project.always.security.oauth.entity.User;
+import com.project.always.security.oauth.dto.request.UserNameRequestDto;
 import com.project.always.security.oauth.jwt.JwtTokenProvider;
-import com.project.always.security.oauth.repository.UserRepository;
 import com.project.always.security.oauth.service.CustomOAuth2UserService;
 import com.project.always.security.oauth.service.UserService;
-import io.restassured.module.mockmvc.RestAssuredMockMvc;
 import java.util.Optional;
-import org.assertj.core.api.Assertions;
-import org.json.JSONObject;
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -40,24 +30,16 @@ import org.springframework.beans.factory.annotation.Autowired;
 import static org.springframework.restdocs.headers.HeaderDocumentation.requestHeaders;
 import static org.springframework.restdocs.headers.HeaderDocumentation.headerWithName;
 
-import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
-import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
-import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.http.HttpHeaders;
-import org.springframework.http.HttpStatus;
 
 import static org.mockito.BDDMockito.given;
 
 import org.springframework.http.MediaType;
-import org.springframework.restdocs.mockmvc.RestDocumentationRequestBuilders;
 import org.springframework.restdocs.payload.JsonFieldType;
 import org.springframework.restdocs.snippet.Snippet;
-import org.springframework.security.test.context.support.WithMockUser;
 import org.springframework.test.web.servlet.ResultActions;
 
-@SpringBootTest
-@AutoConfigureMockMvc
 class OauthControllerApiTest extends AcceptanceTest {
 
     private static final String BEARER_TYPE = "Bearer ";
